update0428

+++ src/main/java/com/ecommerce/entity/BeverageGoods.java
@@ -37,7 +37,7 @@ public class BeverageGoods {
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "BEVERAGE_GOODS_SEQ")
     @SequenceGenerator(name = "BEVERAGE_GOODS_SEQ", sequenceName = "BEVERAGE_GOODS_SEQ", allocationSize = 1)
 	@Column(name = "GOODS_ID")
-	private Long goodsId;
+	private Integer goodsId;
 	
 	@Column(name = "GOODS_NAME")
 	private String goodsName;
